Create `froglight-registry` crate and `SimpleRegistry` SimpleRegistry as two halves, DefaultRegistry and RuntimeRegistry.
The DefaultRegistry is created on startup, and can only be modified by bevy Plugins.

The RuntimeRegistry is created from a DefaultRegistry (creates one if there is none), and is modified by both bevy Plugins and information from servers.

When the client leaves a server, all RuntimeRegistries are overwritten by copies of the corresponding DefaultRegistry.