Create `CameraPlugin`, create 4 cameras for rendering Cameras are layered on top of each other in a way where things that should be on top, are.
The player's hand is rendered with the `ModelCamera`, which will *always* render on top of the `WorldCamera`.

This requires setting the model's `RenderLayers`, or adding the `ModelCameraLayer` component which will manage that for you ( I <3 hooks ).

Same system applies for the `UiCamera` and `OverlayCamera` components. `OverlayCamera` is *always* rendered on top of the `UiCamera`, can can be set with the `OverlayCameraLayer` component.