Create empty crates and import expected dependencies, propagate features
I've got better ideas on how to keep things *modular*, but still easy-ish and straighforward to use.

Blocks in a dyn Block-map is great for supporting custom blocks, but what about registry values?

instead of hardcoding enums, do something similar: In a typed-registry resource, store a Vec of dyn-Trait values.