Implement block upgrading/downgrading where attributes are identical
If the block attributes are identical between versions, then the "indexes" into the attribute array are also identical.

When this is the case you can use `Block::into_version::<Version>()` or just `Block::into()` where the version can be inferred.

It's still possible to convert block with different attributes, but the block type need to be known in order to get the attributes and then each case needs to be handled individually. Not impossible, but generating a match statement for every possible block and handling each case is not a trivial amount of work.

Once written though, it might usable with an API over-top some generalized autoref magic?

I don't think it's possible to allow users to specify what attribute to insert into upgraded blocks, but replacing non-existent blocks from downgrading should be possible. For example, AirBlock, BarrierBlock, or TargetBlock.

Too many thoughts :)