Move `Block` Version converting into it's own trait, flesh out `-world`
Now each block not covered by the attribute-based blanket can get it's own conversion implementation.

`BlockConverter` is a nice const way of bundling a bunch of `BlockConvert` implementation together, as long as you don't need anything inside. Just create it in a `const` converter and extend with the desired versions, no mucking about with traits and functions.