Add trait method to get specific block attributes
Very handy, but not perfect. The way things are planned, when a block has an attribute all possible values must be valid. Instead, two separate enums will be created. One with all valid attributes and another with all possible attributes.

This will be fixed by adding string representations. The block knows what attribute to query based on the input string, and the attribute returns the correct string value.