Manage dependencies, add `extend` methods to registry storage.
Reading over and preparing a new plan for `froglight-registry`.

The original plan was generating structs for each value type and an associated data type. Now, though, the plan is a lot more manual.
Because the generated values can only expect the structures actually used by the game, anything "custom" will automatically fail to parse.

I don't plan on scrapping using structs/enums to store data. Instead, these registries will have to be hand written.
Not ideal, but honestly not that bad. Once the typically used structures are written, it's a game putting the pieces together and adding fields for new versions.

Oh, and I have to write macros for `froglight-nbt`'s `ConvertCase` macro.