Mostly finish generated part of `froglight-item`
The hardest part will be generating the default NBT data of items, but that's for another time.

The only thing left is to write the `ItemProperty` systems, where you can index into items using `ItemProperties` and `ConvertNbt` instead of dealing with raw NBT.

This needs to be done by hand, unfortunately, as the structure is too precise to generate automatically.