Add `SubWorld` for managing custom contained `World`s
Similar to a `SubApp` but has no schedules or runners and lives as a `Component` attached to an `Entity`.

Very powerful, but is quite complex to use efficiently.

For example, I would use this over `SubApp`s when writing a swarm of MC bots.
Each bot doesn't need to manage schedules, ticks, etc, but may want track world state uniquely.
In this case, one `Swarm` `SubApp` (not using the plugin) can handle running systems and each bot can manage it's own `SubWorld`.