Add `FrogPackets` and rewrite and finish `FrogBuf` macros, start packets
`FrogBuf` is so, so much better. Yeah not every part is reusable now, but that made everything 100x harder to read.

Using `FrogReadVersion` and `FrogWriteVersion` for packet enums should help prevent packets being used where they should not be, but it's not a perfect solution.

Ideally, instead of implementing it generically for all versions each packet supports it should be implemented per version, but this adds quite a bit of boilerplate.