Add `no_std` to most crates
There are likely plenty of mistakes that can be fixed in the future, but my work mostly ends here.

Only `froglight-brigader`, `froglight-io`, and `froglight-network` are incompatible!

Brigadier could be made to work, but `petgraph` unfortunately requires `std`.
Bevy fixed this by porting/writing their own, which I am not doing.