Split `froglight-player` into `froglight-text` and `froglight-entity`
Text and chat belongs in its own crate, while player data is part of entity data.

Also considering cutting out the packets from `froglight-network` and making `froglight-packet`,
Then rewriting `froglight-network` as a `no_std` crate for networking over a generic backend,
 of which a `std` one will be provided as part of `froglight-io`.

That would make it possible/easy to integrate proxies, etc. as well. It'll be a lot of work, but it's probably worth it?