Fix chat message signing, support `no_std` better
Salts are per message, not per session. Also spun out the message details into it's own struct for later access.

Add the ability to pass a custom `Duration` fetching implementation.
The standard library has `Instant` and `SystemTime` to get a proper timestamp, but that obviously doesn't work for all platforms.