Start cleaning up `froglight-registry`, prepare for new codegen
Initially this was intended for holding arbitrary data structures, and it still can, but the idea behind it has shifted a little bit.

While some data is hardcoded, those are typically only an index and an identifier.
These can be handled like a HashSet where the "data" is actually a zero-sized unit, and the IndexMap takes care of the index.

Other data is read as JSON from resource packs (or the internal resource pack) and only available at runtime.

Whether to leave this up to the client or provide the defaults in the library while allowing overrides is TBD.
While I would prefer having as little data in code as possible, I'm leaning toward the second solution.