Add the `froglight-utils` crate with premade static storages
This crate will be kept pretty sparse, but it's for stuff not quite general enough for `common` but still used across a few crates.

In this case, the same-ish storage structure is being used across `block`, `entity`, `item`, `registry`, and probably one or two more as crates are written (achievements, statistics, etc?).