Add entity attributes to `froglight-entity`
Well, that was pretty easy.

Now all I have to do is extract the default attributes per entity and add a method to `EntityType` to create an `EntityAttributeSet`.
That's like, 90% of entity stuff completely done!

The rest is parsing NBT from/into components, which is going to be a pain since there's no straightforward method to extract that data.