Add an example for invalid entity types
`SplashPotion` doesn't exist in 1.21.4, therefore it cannot be "spawned".

There is nothing stopping you from inserting the component manually.
Actually, in this case it's probably a good idea to do so, but it must be done *explicitly*.
Since `SplashPotion` cannot be created in 1.21.4, it will never happen automatically.

The same applies the other way around, `Potion` doesn't exist in 1.21.5 and cannot be automatically spawned.
Adding `Potion` may be nice for being able to keep track of all thrown potions, but it must be added *explicitly*.