Add `UnsizedVec` and `UnsizedBuffer`, manually write `BundleDelimiter`s
With the separation of `froglight-io`, `froglight-packet`, and `froglight-network`, it isn't possible (read: I don't want to) to write code to handle the special case of reading a bundle delimiter packet.

So instead, encode it into the protocol: A bundle delimiter is read/written as a bundle of packets.

There are only two real downsides to this approach: having a specific bundle delimiter packet per version, and not having each packet on its own to build and send events. These are just nitpicks, though.